* Migrate Scraper to SDK.V1 * Migrate LegacyCreator data to SDK.V1.Creator

+++ VirtualHole.SDK.V1/DB/Contents/Types/YouTubeVideo.cs
@@ -10,8 +10,8 @@ namespace VirtualHole.DB.Contents
 		public string ThumbnailUrl { get; set; } = string.Empty;
 		public string Description { get; set; } = string.Empty;
 		public TimeSpan Duration { get; set; } = TimeSpan.MinValue;
-		public int ViewsCount { get; set; } = 0;
-		public int LikesCount { get; set; } = 0;
-		public int DislikesCount { get; set; } = 0;
+		public long ViewsCount { get; set; } = 0;
+		public long LikesCount { get; set; } = 0;
+		public long DislikesCount { get; set; } = 0;
 	}
 }
